requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow==10.4.0
python-multipart==0.0.9
httpx==0.27.2
requests==2.32.3
transformers==4.44.2
torch==2.4.0
accelerate==0.34.2
safetensors==0.4.4
vllm==0.6.2
openai==1.45.0
python-dotenv==1.0.1
